requirements.txt
arrow
discord.py
helloasso-apiv5
pyhumps
pyyaml
requests
